requirements-py3.txt
django
requests
pyquery
Pillow
pyBarcode
gunicorn
